--- src/slic3r/plugin/PluginHostApi.hpp
@@ -1,13 +1,0 @@
-#pragma once
-
-#include <pybind11/pybind11.h>
-
-namespace Slic3r {
-
-class PluginHostApi
-{
-public:
-    static void RegisterBindings(pybind11::module_& module);
-};
-
-} // namespace Slic3r
