+++ src/slic3r/plugin/host/PluginHostApp.cpp
@@ -1,0 +1,60 @@
+#include "PluginHostBindings.hpp"
+
+#include <libslic3r/Model.hpp>
+#include <libslic3r/PresetBundle.hpp>
+#include <slic3r/GUI/GUI_App.hpp>
+#include <slic3r/GUI/Plater.hpp>
+
+#include <memory>
+#include <stdexcept>
+
+namespace py = pybind11;
+
+namespace Slic3r {
+namespace {
+
+GUI::Plater* current_plater()
+{
+    if (wxTheApp == nullptr)
+        throw std::runtime_error("OrcaSlicer application is not initialized");
+
+    GUI::Plater* plater = GUI::wxGetApp().plater();
+    if (plater == nullptr)
+        throw std::runtime_error("Plater is not available");
+
+    return plater;
+}
+
+PresetBundle* current_preset_bundle()
+{
+    if (wxTheApp == nullptr)
+        throw std::runtime_error("OrcaSlicer application is not initialized");
+
+    PresetBundle* preset_bundle = GUI::wxGetApp().preset_bundle;
+    if (preset_bundle == nullptr)
+        throw std::runtime_error("Preset bundle is not available");
+
+    return preset_bundle;
+}
+
+} // namespace
+
+// Access to the live GUI application: the Plater and the module-level
+// plater()/model()/preset_bundle() accessors. Everything here is owned by the
+// app and only reachable once the GUI is up (the accessors throw before that).
+void host_bindings::register_app(py::module_& host)
+{
+    py::class_<GUI::Plater, std::unique_ptr<GUI::Plater, py::nodelete>>(host, "Plater")
+        .def("model", static_cast<Model& (GUI::Plater::*)()>(&GUI::Plater::model), py::return_value_policy::reference_internal)
+        .def("is_project_dirty", &GUI::Plater::is_project_dirty)
+        .def("is_presets_dirty", &GUI::Plater::is_presets_dirty)
+        .def("inside_snapshot_capture", &GUI::Plater::inside_snapshot_capture);
+
+    host.def("plater", &current_plater, py::return_value_policy::reference);
+    host.def("model", []() -> Model& {
+        return current_plater()->model();
+    }, py::return_value_policy::reference);
+    host.def("preset_bundle", &current_preset_bundle, py::return_value_policy::reference);
+}
+
+} // namespace Slic3r
